Re-implemented 369448a167e8b3da4ca5bca0b3307500c3371828 in zig
Biggest benefit is the size of static binary: 44KiB (x86_64) when built with:
$ zig build -Doptimize=ReleaseSmall -DPR_SET_CHILD_SUBREAPER=36 -DPR_GET_CHILD_SUBREAPER=37

Other benefits:
* BitField logic (expect_status) is more maintainable
* More implicit error handling.

I have deliberately maintained the code structure (same symbols) so that
it is easy to follow the refactoring.

A minor TODO is to upstream the tiny fix in getopt.zig
(to https://github.com/dmgk/zig-getopt) and move to using
getopt as dependency instead of local fork.